refactor cfg util
