+++ cfgutil/amount.go
