+++ cmd/sweepaccount/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/btcsuite/btcwallet/wallet/txrules"
 	"github.com/btcsuite/btcwallet/wallet/txsizes"
 	"github.com/jessevdk/go-flags"
-	"github.com/stroomnetwork/btcwallet/internal/cfgutil"
+	"github.com/stroomnetwork/btcwallet/cfgutil"
 	"github.com/stroomnetwork/btcwallet/netparams"
 	"github.com/stroomnetwork/btcwallet/wallet/txauthor"
 	"golang.org/x/term"
