+++ rpc/rpcserver/server.go
@@ -33,8 +33,8 @@ import (
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
 	"github.com/btcsuite/btcwallet/walletdb"
+	"github.com/stroomnetwork/btcwallet/cfgutil"
 	"github.com/stroomnetwork/btcwallet/chain"
-	"github.com/stroomnetwork/btcwallet/internal/cfgutil"
 	"github.com/stroomnetwork/btcwallet/internal/zero"
 	"github.com/stroomnetwork/btcwallet/netparams"
 	pb "github.com/stroomnetwork/btcwallet/rpc/walletrpc"
