+++ run/config.go
@@ -20,7 +20,7 @@ import (
 	"github.com/btcsuite/btcd/chaincfg"
 	flags "github.com/jessevdk/go-flags"
 	"github.com/lightninglabs/neutrino"
-	"github.com/stroomnetwork/btcwallet/internal/cfgutil"
+	"github.com/stroomnetwork/btcwallet/cfgutil"
 	"github.com/stroomnetwork/btcwallet/internal/legacy/keystore"
 	"github.com/stroomnetwork/btcwallet/netparams"
 	"github.com/stroomnetwork/btcwallet/wallet"
